Fixed handling of package which resulted in test failure.
The program failed TEST*JSON-SYMBOLS-PACKAGE*.  That was due to the
manner of operation of WITH-OLD-DECODER-SEMANTICS which worked
essentially by supplying a canonical prototype, so that the decoder
has a preconception of what kind of object to create--and where to
intern names.  The last part was clearly wrong: there should be no
such preconception, as the user might be setting *JSON-SYMBOLS-PACKAGE*
(outside WITH-OLD-DECODER-SEMANTICS) to a non-default value, and it is
this value rather than the one supplied in the prototype that should
be obeyed.

darcs-hash:20071231145109-cc086-918066bcfcc2fe6b90f2e89e297bb6fa404fa2e5.gz