catch-read-numeric-overflow-on-allegro
On Allegro Common Lisp, when there's a numeric overflow in the reader,
they don't signal a reader-error, but a simple-error (I have reported this
and they may fix it in the future).  This patch makes use of information
about that simple-error to catch the case in parse-number and treat it
the way parse-number overflows are handled on other lisps.

darcs-hash:20100105201756-b37ea-699f0644f511321c9970c997685f78c64a1ec573.gz