fix-test-after-method-bug-from-juanjo
Ignore-this: 12cec5052070f37a2f72e4e5b0587c32

juanjo (Juan Jose Garcia-Ripoli) found a bug in the binding of the
*JSON-TEST-FILES-PATH* global.  Fixed by moving the initializer into
the test code, and using the ASDF:SYSTEM-RELATIVE-PATHNAME function.

At the same time, killed the testjson.lisp file, which does the testing
as a side-effect of loading.  This is unnecessary because the testing
is done in the PERFORM method of the TEST-OP.

darcs-hash:20121014203847-b37ea-70f64a30a5c7f3ce5e2d4c01c3222ea12ed04bd8.gz